separate streamServerRenderedReactComponent from ReactOnRails
