+++ spec/dummy/config/webpack/alias.js
@@ -4,7 +4,6 @@ module.exports = {
   resolve: {
     alias: {
       Assets: resolve(__dirname, '..', '..', 'client', 'app', 'assets'),
-      stream: 'stream-browserify',
     },
   },
 };
